+++ src/main/java/calculator/Operator.java
@@ -1,8 +1,6 @@
-package step1;
+package calculator;
 
-import java.util.Arrays;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 import java.util.function.BinaryOperator;
