+++ src/main/java/calculator/StringCalculator.java
@@ -1,4 +1,4 @@
-package step1;
+package calculator;
 
 public class StringCalculator {
     private String inputRequest;
