+++ src/main/java/racingCar/domain/Movable.java
@@ -1,4 +1,4 @@
-package retry.domain;
+package racingCar.domain;
 
 @FunctionalInterface
 public interface Movable {
