+++ src/main/java/racingCar/view/InputView.java
@@ -1,4 +1,4 @@
-package retry.view;
+package racingCar.view;
 
 import java.util.Scanner;
