--- src/main/java/step2/Car.java
@@ -1,17 +1,0 @@
-package step2;
-
-public class Car {
-    private int position;
-
-    public Car(int position) {
-        this.position = position;
-    }
-
-    public void setPosition(int position) {
-        this.position = position;
-    }
-
-    public int getPosition() {
-        return position;
-    }
-}
