+++ src/test/java/study/calculator/StringCalculatorTest.java
@@ -1,12 +1,11 @@
-package study.step1;
+package study.calculator;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.EmptySource;
 import org.junit.jupiter.params.provider.NullSource;
 import org.junit.jupiter.params.provider.ValueSource;
-import step1.StringCalculator;
+import calculator.StringCalculator;
 
 import static org.assertj.core.api.Assertions.*;
 public class StringCalculatorTest {
