+++ src/test/java/study/racingcar/CarsTest.java
@@ -1,9 +1,9 @@
-package study.retry;
+package study.racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import retry.domain.Car;
-import retry.domain.Cars;
+import racingCar.domain.Car;
+import racingCar.domain.Cars;
 
 import java.util.Arrays;
 import java.util.List;
