+++ src/test/java/study/racingcar/RacingGameTest.java
@@ -1,10 +1,10 @@
-package study.retry;
+package study.racingcar;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import retry.domain.Car;
-import retry.domain.Cars;
-import retry.domain.RacingGame;
+import racingCar.domain.Car;
+import racingCar.domain.Cars;
+import racingCar.domain.RacingGame;
 
 import java.util.Arrays;
 import java.util.List;
