--- src/test/java/study/step3/RacingGameTestDoc.md
@@ -1,14 +1,0 @@
-주어진 횟수 동안 n대의 자동차는 전진 or 멈출 수 있다.
-
-사용자는 몇 대의 자동차로 몇 번의 이동을 할 것인가
-
-자동차 이름은 쉼표(,)를 기준으로 구분
-
-자동차 경주 게임을 완료한 후 누가 우승했는지를 알려준다. 우승자는 한명 이상일 수 있다.
-
-전진 조건 0~9 사이 랜덤 값에서 4 이상인 경우
-
-MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, 
-
-domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현..
-
